+++ CloudShell.Providers.Applications/ApplicationResourceService.Declarations.cs
@@ -14,23 +14,6 @@ public sealed partial class ApplicationResourceService
         .Select(ResolveDefinition)
         .ToArray();
 
-    public IReadOnlyList<ApplicationContainerDeployment> GetContainerDeployments(string applicationId) =>
-        containerDeployments.List(applicationId);
-
-    public IReadOnlyList<ApplicationContainerRevisionHistoryEntry> GetContainerRevisions(string applicationId)
-    {
-        var revisions = containerDeployments.ListRevisions(applicationId);
-        if (revisions.Count > 0)
-        {
-            return AssignContainerRevisionHistoryNumbers(revisions);
-        }
-
-        var application = GetApplication(applicationId);
-        return application is null
-            ? []
-            : AssignContainerRevisionHistoryNumbers(CreateContainerRevisionHistoryEntries(application));
-    }
-
     public bool CanApplyDeclaration(ResourceDeclaration declaration) =>
         ApplicationResourceProviderIds.IsApplicationProvider(declaration.ProviderId);